// src/modules/trades/dto/pair-coin.dto.ts
export class PairCoin {
  pairs_id: number;
  quote_id: number;
  base_id: number;
  quote_name: string;
  base_name: string;
  base_symbol: string;
  base_type: number;
  base_stellar_issuer: string;
  base_bsc_address: string;
  base_decimal: number;
  quote_symbol: string;
  quote_type: number;
  quote_stellar_issuer: string;
  quote_bsc_address: string;
  quote_decimal: number;
  price_precision: string;
}